revert refactor of resolve_path method
