+++ src/resource_reader.rs
@@ -1,6 +1,4 @@
-
-use std::path::Path;
-use futures_util::{TryFutureExt, future::join_all};
+use futures_util::{future::join_all, TryFutureExt};
 use url::Url;
 
 use crate::runtime::TargetRuntime;
@@ -57,41 +55,5 @@ impl ResourceReader {
             .into_iter()
             .collect::<anyhow::Result<Vec<_>>>()?;
         Ok(content)
-    }
-
-    /// Checks if path is absolute else it joins file path with relative dir
-    /// path
-    pub fn resolve_path(src: &str, root_dir: Option<&Path>) -> String {
-        if Path::new(&src).is_absolute() {
-            src.to_string()
-        } else {
-            let path = root_dir.unwrap_or(Path::new(""));
-            path.join(src).to_string_lossy().to_string()
-        }
-    }
-
-}
-
-
-
-#[cfg(test)]
-mod resource_reader_tests {
-    use std::path::{Path, PathBuf};
-    use pretty_assertions::assert_eq;
-    use crate::resource_reader::ResourceReader;
-    
-    #[test]
-    fn test_relative_path() {
-        let path_dir = Path::new("abc/xyz");
-        let file_relative = "foo/bar/my.proto";
-        let file_absolute = "/foo/bar/my.proto";
-        assert_eq!(
-            path_dir.to_path_buf().join(file_relative),
-            PathBuf::from(ResourceReader::resolve_path(file_relative, Some(path_dir)))
-        );
-        assert_eq!(
-            "/foo/bar/my.proto",
-            ResourceReader::resolve_path(file_absolute, Some(path_dir))
-        );
     }
 }
